+++ x/evm/statedb/native.go
@@ -1,14 +1,14 @@
 package statedb
 
 import (
-	"cosmossdk.io/store/types"
+	"cosmossdk.io/store/cachemulti"
 	"github.com/ethereum/go-ethereum/common"
 )
 
 var _ JournalEntry = nativeChange{}
 
 type nativeChange struct {
-	snapshot types.MultiStore
+	snapshot cachemulti.Store
 	events   int
 }
 
@@ -17,6 +17,6 @@ func (native nativeChange) Dirtied() *common.Address {
 }
 
 func (native nativeChange) Revert(s *StateDB) {
-	s.restoreNativeState(native.snapshot)
+	s.revertNativeStateToSnapshot(native.snapshot)
 	s.nativeEvents = s.nativeEvents[:len(s.nativeEvents)-native.events]
 }